requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.7.0
blis==0.7.10
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.6
cmake==3.27.2
coloredlogs==15.0.1
colorlog==4.8.0
confection==0.1.1
contourpy==1.1.0
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
dacite==1.8.1
dataclasses-json==0.5.13
emoji==2.8.0
exceptiongroup==1.1.2
fake-useragent==1.2.1
fastapi==0.99.1
filelock==3.12.2
filetype==1.2.0
flake8==4.0.1
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
ginza==5.1.2
ginza-transformers==0.4.2
gitdb==4.0.10
GitPython==3.1.32
google-ads==21.3.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
html2text==2020.1.16
httplib2==0.22.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
isort==5.10.1
ja-ginza-electra==5.1.2
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
langchain==0.0.240
langcodes==3.3.0
langsmith==0.0.14
lit==16.0.6
llama-index==0.7.11.post1
logzero==1.7.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mccabe==0.6.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy==0.910
mypy-extensions==0.4.4
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.1
pandas==2.0.3
parse==1.19.1
pathspec==0.11.1
pathy==0.10.2
Pillow==10.0.0
plac==1.3.5
platformdirs==3.9.1
pluggy==1.2.0
posthog==3.0.2
preshed==3.0.8
proto-plus==1.22.3
protobuf==4.23.4
pulsar-client==3.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pydantic==1.10.12
pyee==8.2.2
pyflakes==2.4.0
pyparsing==3.0.9
PyPika==0.48.9
pyppeteer==1.0.2
pyquery==2.0.0
pysen==0.10.5
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.4.4
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.1.9
SQLAlchemy==2.0.19
srsly==2.4.7
starlette==0.27.0
SudachiDict-core==20230711
SudachiPy==0.6.7
SudachiTra==0.1.8
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
thinc==8.1.12
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==1.2.3
tomlkit==0.11.8
torch==2.0.1
tqdm==4.65.0
transformers==4.25.1
triton==2.0.0
typer==0.7.0
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unidiff==0.7.5
unstructured==0.10.4
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
w3lib==2.1.2
wasabi==0.10.1
watchfiles==0.19.0
websockets==10.4
wikipedia==1.4.0
Wikipedia-API==0.6.0
yarl==1.9.2
zipp==3.16.2
